Identified an error in the Signature generation and verification. For some messages the PKCS encoded 156 byte string was larger than the modulus, which is not allowed by the Montgomery Multiplication algorithm. The problem arises because we put the hash at the end of this string, which means at the MOST SIGNIFICANT position (as all montgomery mathematics was implemented in LITTLE ENDIAN). This was fixed by internally 'turning the number around' from BIG to LITTLE ENDIAN. This is all internal and does not affect anything, but it ensures correct operation because now the most significant byte of every PKCS encoded string is 0x00, so the total number is ALWAYS smaller than the modulus.
Secondly, I have not found and efficient way to calculate rmodn and r2modn (needed for exponentiation, changes per modulus). So it is now required to pass them with the function call to MonExp, and by extension also to the SIGN and VERIFY methods. For RSA, these numbers are provided in keys.c, for DiffieHellman, the number can be found in the example dh_test.c.